Compute usage statistics in a Celery worker
Instead of Flask to avoid restarts.
