+++ packit_service/service/api/usage.py
@@ -1,15 +1,14 @@
 # Copyright Contributors to the Packit project.
 # SPDX-License-Identifier: MIT
 
-from collections import OrderedDict
-from datetime import datetime, timedelta, timezone
+import time
+
+from datetime import datetime, timezone
 from http import HTTPStatus
 from logging import getLogger
-from typing import Any, Union
+from typing import Any
 
-from cachetools.func import ttl_cache
-
-from flask import request, escape
+from flask import request, escape, redirect, Response
 from flask_restx import Namespace, Resource
 
 from packit_service.models import (
@@ -21,27 +20,26 @@ from packit_service.models import (
     SyncReleaseModel,
     TFTTestRunGroupModel,
     VMImageBuildTargetModel,
-    BodhiUpdateTargetModel,
-    KojiBuildTargetModel,
-    SyncReleaseTargetModel,
     get_usage_data,
 )
 from packit_service.service.api.utils import response_maker
 from packit_service.constants import (
-    USAGE_CURRENT_DATE,
-    USAGE_DATE_IN_THE_PAST,
     USAGE_DATE_IN_THE_PAST_STR,
     USAGE_PAST_DAY_DATE_STR,
     USAGE_PAST_WEEK_DATE_STR,
     USAGE_PAST_MONTH_DATE_STR,
     USAGE_PAST_YEAR_DATE_STR,
 )
+from packit_service.celerizer import celery_app
+from packit_service.service.tasks import (
+    get_usage_interval_data,
+    get_past_usage_data,
+    calculate_onboarded_projects,
+)
 
 logger = getLogger("packit_service")
 
 usage_ns = Namespace("usage", description="Data about Packit usage")
-
-_CACHE_MAXSIZE = 100  # can it be removed, should things being cached already lower?
 
 
 @usage_ns.route("")
@@ -345,50 +343,6 @@ class Onboarded2024Q1(Resource):
         HTTPStatus.OK,
         "Onboarded projects for which exist a Bodhi update or a Koji build or a Packit merged PR.",
     )
-    @classmethod
-    def calculate(cls):
-        known_onboarded_projects = (
-            GitProjectModel.get_known_onboarded_downstream_projects()
-        )
-
-        bodhi_updates = BodhiUpdateTargetModel.get_all_projects()
-        koji_builds = KojiBuildTargetModel.get_all_projects()
-        onboarded_projects = bodhi_updates.union(koji_builds).union(
-            known_onboarded_projects
-        )
-
-        # find **downstream git projects** with a PR created by Packit
-        downstream_synced_projects = (
-            SyncReleaseTargetModel.get_all_downstream_projects()
-        )
-        # if there exist a downstream Packit PR we are not sure it has been
-        # merged, the project is *almost onboarded* until the PR is merged
-        # (unless we already know it has a koji build or bodhi update, then
-        # we don't need to check for a merged PR - it obviously has one)
-        almost_onboarded_projects = downstream_synced_projects.difference(
-            onboarded_projects
-        )
-        # do not re-check projects we already checked and we know they
-        # have a merged Packit PR
-        recheck_if_onboarded = almost_onboarded_projects.difference(
-            known_onboarded_projects
-        )
-
-        onboarded = {
-            project.id: project.project_url
-            for project in onboarded_projects.union(known_onboarded_projects)
-        }
-        almost_onboarded = {
-            project.id: project.project_url
-            for project in recheck_if_onboarded.difference(onboarded_projects)
-        }
-
-        return {"onboarded": onboarded, "almost_onboarded": almost_onboarded}
-
-    @classmethod
-    def get_num_of_onboarded_projects(cls):
-        return len(cls.calculate()["onboarded"])
-
     def get(self):
         """
         Returns a list of onboarded projects for which exist at least a
@@ -402,193 +356,124 @@ class Onboarded2024Q1(Resource):
         Examples:
         /api/usage/onboarded-projects
         """
-        return self.calculate()
+        return calculate_onboarded_projects()
 
 
-def _get_past_usage_data(datetime_from=None, datetime_to=None, top=5):
-    # Even though frontend expects only the first N (=5) to be present
-    # in the project lists, we need to get all to calculate the number
-    # of active projects.
-    # (This info will be added to the payload for frontend.)
-    # The original `top` argument will be used later
-    # to get the expected number of projects in the response.
-    top_all_project = 100000
+def _get_celery_result(id: str) -> Response:
+    """
+    Present the Celery task result.
 
-    raw_result = get_usage_data(
-        datetime_from=datetime_from, datetime_to=datetime_to, top=top_all_project
-    )
-    return response_maker(
-        {
-            "active_projects": raw_result["active_projects"],
-            "jobs": {
-                job: {
-                    "job_runs": data["job_runs"],
-                    "top_projects_by_job_runs": dict(
-                        list(OrderedDict(data["top_projects_by_job_runs"]).items())[
-                            :top
-                        ]
-                    ),
-                    "active_projects": len(data["top_projects_by_job_runs"]),
-                }
-                for job, data in raw_result["jobs"].items()
-            },
-            "onboarded_projects_q1_2024": Onboarded2024Q1.get_num_of_onboarded_projects(),
-        }
-    )
+    The redirect link provided by the below api functions
+    is meant to be polled by the UX.
+
+    Wait here until the UX can deal with polling for the result.
+    """
+    TIMEOUT = 5  # seconds
+    STEP = 0.1  # second
+    elapsed = 0.0
+    while not (celery_app.AsyncResult(id).ready() or elapsed > TIMEOUT):
+        elapsed += STEP
+        time.sleep(STEP)
+    result = celery_app.AsyncResult(id)
+    return response_maker(result.result)
 
 
 @usage_ns.route("/past-day")
 class UsagePastDay(Resource):
-    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit usage")
-    @ttl_cache(maxsize=_CACHE_MAXSIZE, ttl=timedelta(hours=1).total_seconds())
+    @usage_ns.response(
+        HTTPStatus.OK, "Provides a url where to wait for Packit last day usage"
+    )
     def get(self):
-        return _get_past_usage_data(datetime_from=USAGE_PAST_DAY_DATE_STR)
+        task = get_past_usage_data.delay(datetime_from=USAGE_PAST_DAY_DATE_STR)
+        return _get_celery_result(task.id)
+
+
+@usage_ns.route("/past-day/<id>")
+@usage_ns.param("id", "Celery task id")
+class UsagePastDayResult(Resource):
+    @usage_ns.response(HTTPStatus.OK, "Provide data about Packit last day usage")
+    def get(self, id):
+        return _get_celery_result(id)
 
 
 @usage_ns.route("/past-week")
 class UsagePastWeek(Resource):
-    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit usage")
-    @ttl_cache(maxsize=_CACHE_MAXSIZE, ttl=timedelta(hours=1).total_seconds())
+    @usage_ns.response(
+        HTTPStatus.OK, "Provides a url where to wait for Packit last week usage"
+    )
     def get(self):
-        return _get_past_usage_data(datetime_from=USAGE_PAST_WEEK_DATE_STR)
+        task = get_past_usage_data.delay(datetime_from=USAGE_PAST_WEEK_DATE_STR)
+        return redirect(f"past-week/{task.id}", code=302)
+
+
+@usage_ns.route("/past-week/<id>")
+@usage_ns.param("id", "Celery task id")
+class UsagePastWeekResult(Resource):
+    @usage_ns.response(HTTPStatus.OK, "Provide data about Packit last week usage")
+    def get(self, id):
+        return _get_celery_result(id)
 
 
 @usage_ns.route("/past-month")
 class UsagePastMonth(Resource):
-    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit usage")
-    @ttl_cache(maxsize=_CACHE_MAXSIZE, ttl=timedelta(days=1).total_seconds())
+    @usage_ns.response(
+        HTTPStatus.OK, "Provides a url where to wait for Packit last month usage"
+    )
     def get(self):
-        return _get_past_usage_data(datetime_from=USAGE_PAST_MONTH_DATE_STR)
+        task = get_past_usage_data.delay(datetime_from=USAGE_PAST_MONTH_DATE_STR)
+        return redirect(f"past-month/{task.id}", code=302)
+
+
+@usage_ns.route("/past-month/<id>")
+@usage_ns.param("id", "Celery task id")
+class UsagePastMonthResult(Resource):
+    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit last month usage")
+    def get(self, id):
+        return _get_celery_result(id)
 
 
 @usage_ns.route("/past-year")
 class UsagePastYear(Resource):
-    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit usage")
-    @ttl_cache(maxsize=_CACHE_MAXSIZE, ttl=timedelta(days=1).total_seconds())
+    @usage_ns.response(
+        HTTPStatus.OK, "Provides a url where to wait for Packit last year usage"
+    )
     def get(self):
-        return _get_past_usage_data(datetime_from=USAGE_PAST_YEAR_DATE_STR)
+        task = get_past_usage_data.delay(datetime_from=USAGE_PAST_YEAR_DATE_STR)
+        return redirect(f"past-year/{task.id}", code=302)
+
+
+@usage_ns.route("/past-year/<id>")
+@usage_ns.param("id", "Celery task id")
+class UsagePastYearResult(Resource):
+    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit last year usage")
+    def get(self, id):
+        return _get_celery_result(id)
 
 
 @usage_ns.route("/total")
 class UsageTotal(Resource):
-    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit usage")
-    @ttl_cache(maxsize=_CACHE_MAXSIZE, ttl=timedelta(days=1).total_seconds())
+    @usage_ns.response(
+        HTTPStatus.OK, "Provides a url where to wait for Packit total usage data"
+    )
     def get(self):
-        return _get_past_usage_data(datetime_from=USAGE_DATE_IN_THE_PAST_STR)
+        task = get_past_usage_data.delay(datetime_from=USAGE_DATE_IN_THE_PAST_STR)
+        return redirect(f"total/{task.id}", code=302)
 
 
-# format the chart needs is a list of {"x": "datetimelegend", "y": value}
-CHART_DATA_TYPE = list[dict[str, Union[str, int]]]
-
-
-@ttl_cache(maxsize=_CACHE_MAXSIZE, ttl=timedelta(hours=1).total_seconds())
-def _get_usage_interval_data(
-    days: int, hours: int, count: int
-) -> dict[str, Union[str, CHART_DATA_TYPE, dict[str, CHART_DATA_TYPE]]]:
-    """
-    :param days: number of days for the interval length
-    :param hours: number of days for the interval length
-    :param count: number of intervals
-    :return: usage data for the COUNT number of intervals
-      (delta is DAYS number of days and HOURS number of hours)
-    """
-    delta = timedelta(days=days, hours=hours)
-
-    current_date = USAGE_CURRENT_DATE
-    days_legend = []
-    for _ in range(count):
-        days_legend.append(current_date)
-        current_date -= delta
-
-    result_jobs: dict[str, CHART_DATA_TYPE] = {}
-    result_jobs_project_count: dict[str, CHART_DATA_TYPE] = {}
-    result_jobs_project_cumulative_count: dict[str, CHART_DATA_TYPE] = {}
-    result_events: dict[str, CHART_DATA_TYPE] = {}
-    result_active_projects: CHART_DATA_TYPE = []
-    result_active_projects_cumulative: CHART_DATA_TYPE = []
-
-    logger.warn(
-        f"Getting usage data datetime_from {USAGE_DATE_IN_THE_PAST} datetime_to {days_legend[-1]}"
-    )
-    past_data = get_usage_data(
-        datetime_from=USAGE_DATE_IN_THE_PAST, datetime_to=days_legend[-1], top=100000
-    )
-    logger.warn("Got usage data ")
-    cumulative_projects_past = set(
-        past_data["active_projects"]["top_projects_by_events_handled"].keys()
-    )
-    cumulative_projects = cumulative_projects_past.copy()
-    cumulative_projects_for_jobs_past = {
-        job: set(data["top_projects_by_job_runs"].keys())
-        for job, data in past_data["jobs"].items()
-    }
-    cumulative_projects_for_jobs = cumulative_projects_for_jobs_past.copy()
-
-    for day in reversed(days_legend):
-        day_from = (day - delta).isoformat()
-        day_to = day.isoformat()
-        legend = day.strftime("%H:%M" if (hours and not days) else "%Y-%m-%d")
-
-        interval_result = get_usage_data(
-            datetime_from=day_from, datetime_to=day_to, top=100000
-        )
-
-        for job, data in interval_result["jobs"].items():
-            result_jobs.setdefault(job, [])
-            result_jobs[job].append({"x": legend, "y": data["job_runs"]})
-            result_jobs_project_count.setdefault(job, [])
-            result_jobs_project_count[job].append(
-                {"x": legend, "y": len(data["top_projects_by_job_runs"])}
-            )
-
-            cumulative_projects_for_jobs[job] |= data["top_projects_by_job_runs"].keys()
-            result_jobs_project_cumulative_count.setdefault(job, [])
-            result_jobs_project_cumulative_count[job].append(
-                {"x": legend, "y": len(cumulative_projects_for_jobs[job])}
-            )
-
-        for event, data in interval_result["events"].items():
-            result_events.setdefault(event, [])
-            result_events[event].append({"x": legend, "y": data["events_handled"]})
-
-        result_active_projects.append(
-            {"x": legend, "y": interval_result["active_projects"].get("project_count")}
-        )
-        cumulative_projects |= interval_result["active_projects"][
-            "top_projects_by_events_handled"
-        ].keys()
-        result_active_projects_cumulative.append(
-            {"x": legend, "y": len(cumulative_projects)}
-        )
-
-    onboarded_projects_per_job = {}
-    for job, data in past_data["jobs"].items():
-        onboarded_projects_per_job[job] = list(
-            cumulative_projects_for_jobs[job] - cumulative_projects_for_jobs_past[job]
-        )
-
-    return response_maker(
-        {
-            "jobs": result_jobs,
-            "jobs_project_count": result_jobs_project_count,
-            "jobs_project_cumulative_count": result_jobs_project_cumulative_count,
-            "events": result_events,
-            "from": days_legend[0].isoformat(),
-            "to": days_legend[-1].isoformat(),
-            "active_projects": result_active_projects,
-            "active_projects_cumulative": result_active_projects_cumulative,
-            "onboarded_projects": list(cumulative_projects - cumulative_projects_past),
-            "onboarded_projects_per_job": onboarded_projects_per_job,
-        }
-    )
+@usage_ns.route("/total/<id>")
+@usage_ns.param("id", "Celery task id")
+class UsageTotalResult(Resource):
+    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit total usage")
+    def get(self, id):
+        return _get_celery_result(id)
 
 
 @usage_ns.route("/intervals")
 class UsageIntervals(Resource):
-    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit usage")
+    @usage_ns.response(HTTPStatus.OK, "Ask data about Packit interval usage")
     def get(self):
         """
-        Returns the data for trend charts.
+        Returns a new url where to wait for Celery task results.
 
         Use `days` and `hours` parameters to define interval and `count` to set number of intervals.
 
@@ -599,4 +484,18 @@ class UsageIntervals(Resource):
         count = int(escape(request.args.get("count", "10")))
         delta_hours = int(escape(request.args.get("hours", "0")))
         delta_days = int(escape(request.args.get("days", "0")))
-        return _get_usage_interval_data(hours=delta_hours, days=delta_days, count=count)
+        task = get_usage_interval_data.delay(
+            hours=delta_hours, days=delta_days, count=count
+        )
+        return redirect(f"intervals/{task.id}", code=302)
+
+
+@usage_ns.route("/intervals/<id>")
+@usage_ns.param("id", "Celery task id")
+class UsageIntervalsResult(Resource):
+    @usage_ns.response(HTTPStatus.OK, "Providing data about Packit usage")
+    def get(self, id):
+        """
+        Returns the data for trend charts collected by a celery worker.
+        """
+        return _get_celery_result(id)
